Modified ipam allocation
