+++ examples/vl3_basic/vl3-nse/cmd/nse.go
@@ -16,28 +16,17 @@
 package main
 
 import (
-	"bytes"
 	"context"
 	"flag"
-	"fmt"
-	"github.com/networkservicemesh/examples/examples/universal-cnf/vppagent/pkg/ucnf"
-	"github.com/networkservicemesh/networkservicemesh/controlplane/api/networkservice"
-	"golang.org/x/sync/errgroup"
-	"os"
-	"strings"
-	"sync"
-	"time"
-
-	"google.golang.org/grpc"
-
 	"github.com/danielvladco/k8s-vnet/pkg/nseconfig"
-	"github.com/dtornow/cnns-nsr/cnns-ipam/api/ipprovider"
-	"github.com/gofrs/uuid"
+	"github.com/networkservicemesh/examples/examples/universal-cnf/vppagent/pkg/ucnf"
 	"github.com/networkservicemesh/examples/examples/universal-cnf/vppagent/pkg/vppagent"
+	"github.com/networkservicemesh/networkservicemesh/controlplane/api/networkservice"
 	"github.com/networkservicemesh/networkservicemesh/pkg/tools"
 	"github.com/networkservicemesh/networkservicemesh/sdk/common"
-	"github.com/networkservicemesh/networkservicemesh/sdk/endpoint"
 	"github.com/sirupsen/logrus"
+	"os"
+	"strings"
 )
 
 const (
@@ -61,80 +50,11 @@ func (mf *Flags) Process() {
 }
 
 type vL3CompositeEndpoint struct {
-	IpamAllocator     ipprovider.AllocatorClient
-	registeredSubnets chan *ipprovider.Subnet
-	mu                *sync.RWMutex
-}
-
-func (e vL3CompositeEndpoint) Cleanup(ctx context.Context) error {
-	var errs errors
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	for s := range e.registeredSubnets {
-		_, err := e.IpamAllocator.FreeSubnet(ctx, s)
-		if err != nil {
-			errs = append(errs, err)
-		}
-	}
-	if len(errs) > 0 {
-		return errs
-	}
-	return nil
-}
-
-func (e vL3CompositeEndpoint) Renew(ctx context.Context, errorHandler func(err error)) error {
-	g, ctx := errgroup.WithContext(ctx)
-	for subnet := range e.registeredSubnets {
-		subnet := subnet
-		g.Go(func() error {
-			for range time.Tick(time.Duration(subnet.LeaseTimeout-1) * time.Hour) {
-				_, err := e.IpamAllocator.RenewSubnetLease(ctx, subnet)
-				if err != nil {
-					errorHandler(err)
-				}
-			}
-			return nil
-		})
-	}
-	return g.Wait()
 }
 
 func (e vL3CompositeEndpoint) AddCompositeEndpoints(nsConfig *common.NSConfiguration, ucnfEndpoint *nseconfig.Endpoint) *[]networkservice.NetworkServiceServer {
-	var subnet *ipprovider.Subnet
-	for i := 0; i < 6; i++ {
-		var err error
-		subnet, err = e.IpamAllocator.AllocateSubnet(context.Background(), &ipprovider.SubnetRequest{
-			Identifier: &ipprovider.Identifier{
-				Fqdn:               ucnfEndpoint.CNNS.Address,
-				Name:               ucnfEndpoint.CNNS.Name + uuid.Must(uuid.NewV4()).String(),
-				ConnectivityDomain: ucnfEndpoint.CNNS.ConnectivityDomain,
-			},
-			AddrFamily: &ipprovider.IpFamily{Family: ipprovider.IpFamily_IPV4},
-			PrefixLen:  uint32(ucnfEndpoint.VL3.IPAM.PrefixLength),
-		})
-		if err != nil {
-			if i == 5 {
-				logrus.Fatal("ipam allocation not successful: ", err)
-			}
-			logrus.Errorf("ipam allocation not successful: %v \n waiting 60 seconds before retrying \n", err)
-			time.Sleep(60 * time.Second)
-		} else {
-			break
-		}
-	}
-	e.mu.Lock()
-	e.registeredSubnets <- subnet
-	e.mu.Unlock()
 
-	prefixPool := subnet.Prefix.Subnet
-
-	logrus.WithFields(logrus.Fields{
-		"prefixPool":         prefixPool,
-		"nsConfig.IPAddress": nsConfig.IPAddress,
-	}).Infof("Creating vL3 IPAM endpoint")
-	ipamEp := endpoint.NewIpamEndpoint(&common.NSConfiguration{
-		IPAddress: prefixPool,
-	})
+	logrus.WithFields(logrus.Fields{}).Infof("Creating vL3 IPAM endpoint")
 
 	var nsRemoteIpList []string
 	nsRemoteIpListStr, ok := os.LookupEnv("NSM_REMOTE_NS_IP_LIST")
@@ -142,8 +62,7 @@ func (e vL3CompositeEndpoint) AddCompositeEndpoints(nsConfig *common.NSConfigura
 		nsRemoteIpList = strings.Split(nsRemoteIpListStr, ",")
 	}
 	compositeEndpoints := []networkservice.NetworkServiceServer{
-		ipamEp,
-		newVL3ConnectComposite(nsConfig, prefixPool,
+		newVL3ConnectComposite(nsConfig, nsConfig.IPAddress,
 			&vppagent.UniversalCNFVPPAgentBackend{}, nsRemoteIpList, func() string {
 				return ucnfEndpoint.NseName
 			}),
@@ -164,60 +83,14 @@ func main() {
 	mainFlags := &Flags{}
 	mainFlags.Process()
 
-	ipamAddress, ok := os.LookupEnv("IPAM_ADDRESS")
-	if !ok {
-		ipamAddress = "cnns-ipam:50051"
-	}
-	conn, err := grpc.Dial(ipamAddress, grpc.WithInsecure())
-	if err != nil {
-		logrus.Fatal("unable to connect to ipam server", err)
-	}
-
-	ipamAllocator := ipprovider.NewAllocatorClient(conn)
-	vl3 := vL3CompositeEndpoint{
-		IpamAllocator:     ipamAllocator,
-		registeredSubnets: make(chan *ipprovider.Subnet),
-		mu:                &sync.RWMutex{},
-	}
-
+	vl3 := vL3CompositeEndpoint{}
 	ctx, cancel := context.WithCancel(context.Background())
-
-	go func() {
-		logrus.Info("begin the renew process")
-		if err := vl3.Renew(ctx, func(err error) {
-			if err != nil {
-				logrus.Error("unable to renew the subnet", err)
-			}
-		}); err != nil {
-			logrus.Error(err)
-		}
-	}()
-
-	ucnfNse := ucnf.NewUcnfNse(mainFlags.ConfigPath, mainFlags.Verify, &vppagent.UniversalCNFVPPAgentBackend{}, vl3)
+	defer cancel()
+	ucnfNse := ucnf.NewUcnfNse(mainFlags.ConfigPath, mainFlags.Verify, &vppagent.UniversalCNFVPPAgentBackend{}, vl3, &ctx)
 	logrus.Info("endpoint started")
 
-	select {
-	case _ = <-c:
-		logrus.Info("cleaning up")
-		if err := vl3.Cleanup(ctx); err != nil {
-			logrus.Error(err)
-		}
-		ucnfNse.Cleanup()
-		logrus.Info("cleaning up complete")
-		cancel()
-		conn.Close()
-	}
-
-}
-
-type errors []error
-
-func (es errors) Error() string {
-	buff := bytes.NewBufferString("multiple errors: \n")
-	for _, e := range es {
-		_, _ = fmt.Fprintf(buff, "\t%s\n", e)
-	}
-	return buff.String()
+	defer ucnfNse.Cleanup()
+	<-c
 }
 
 /*
